+++ phpmyfaq/admin/assets/src/utils/index.ts
@@ -1,3 +1,4 @@
+export * from './flesch-reading-ease';
 export * from './session';
 export * from './sidebar';
 export * from './utils';